requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.2
django-allauth==65.7.0
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modeltranslation==0.19.13
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
idna==3.10
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
whitenoise==6.6.0
